+++ simple-stack/src/test/java/com/zhuinden/simplestack/FlowTest.java
@@ -21,6 +21,8 @@ import android.os.Parcel;
 import android.os.Parcelable;
 import android.support.annotation.NonNull;
 
+import com.zhuinden.simplestack.helpers.TestKey;
+
 import org.junit.Before;
 import org.junit.Test;