+++ simple-stack/src/test/java/com/zhuinden/simplestack/helpers/HasServices.java
@@ -1,0 +1,9 @@
+package com.zhuinden.simplestack.helpers;
+
+import com.zhuinden.simplestack.ScopeKey;
+import com.zhuinden.simplestack.ServiceBinder;
+
+public interface HasServices
+        extends ScopeKey {
+    void bindServices(ServiceBinder serviceBinder);
+}