+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+-Simple Stack 2.0.1 (2019-05-20)
+--------------------------------
+
+- ADDED: `ServiceBinder.addAlias()` to allow adding an alias to an added service: available for look-up, but without getting callbacks by this registration (thus avoiding unnecessary `toBundle()` calls for a multi-registered service).
+
+This also adds `ServiceBinder.hasAlias()`, `Backstack.hasAlias()`, `BackstackDelegate.hasAlias()`, and `Navigator.hasAlias()`.
+
 -Simple Stack 2.0.0 (2019-04-30)
 --------------------------------